IDAM-167 - adding files for deployment
