+++ custom-pages/clients/API Explorer Application.json
@@ -1,0 +1,22 @@
+{
+  "is_token_endpoint_ip_header_trusted": false,
+  "name": "API Explorer Application",
+  "description": "Test Application for API Explorer.",
+  "is_first_party": true,
+  "sso_disabled": false,
+  "cross_origin_auth": false,
+  "oidc_conformant": false,
+  "client_secret": "",
+  "jwt_configuration": {
+    "lifetime_in_seconds": 36000,
+    "secret_encoded": false
+  },
+  "app_type": "non_interactive",
+  "grant_types": [
+    "authorization_code",
+    "implicit",
+    "refresh_token",
+    "client_credentials"
+  ],
+  "custom_login_page_on": true
+}
