+++ custom-pages/clients/OctopusDeploy-HostedPages.json
@@ -1,0 +1,31 @@
+{
+  "is_token_endpoint_ip_header_trusted": false,
+  "name": "OctopusDeploy-HostedPages",
+  "is_first_party": true,
+  "oidc_conformant": true,
+  "sso_disabled": false,
+  "cross_origin_auth": false,
+  "description": "",
+  "logo_uri": "",
+  "sso": false,
+  "callbacks": [],
+  "allowed_logout_urls": [],
+  "allowed_clients": [],
+  "allowed_origins": [],
+  "client_secret": "",
+  "jwt_configuration": {
+    "alg": "RS256",
+    "lifetime_in_seconds": 36000,
+    "secret_encoded": false
+  },
+  "token_endpoint_auth_method": "client_secret_post",
+  "app_type": "non_interactive",
+  "grant_types": [
+    "authorization_code",
+    "implicit",
+    "refresh_token",
+    "client_credentials"
+  ],
+  "web_origins": [],
+  "custom_login_page_on": true
+}
