+++ custom-pages/connections/google-oauth2.json
@@ -1,0 +1,47 @@
+{
+  "options": {
+    "email": true,
+    "profile": true,
+    "scope": [
+      "email",
+      "profile"
+    ],
+    "client_id": "",
+    "client_secret": "",
+    "allowed_audiences": "",
+    "contacts": false,
+    "blogger": false,
+    "calendar": false,
+    "gmail": false,
+    "google_plus": false,
+    "orkut": false,
+    "picasa_web": false,
+    "tasks": false,
+    "youtube": false,
+    "adsense_management": false,
+    "google_affiliate_network": false,
+    "analytics": false,
+    "google_books": false,
+    "google_cloud_storage": false,
+    "content_api_for_shopping": false,
+    "chrome_web_store": false,
+    "document_list": false,
+    "google_drive": false,
+    "google_drive_files": false,
+    "latitude_best": false,
+    "latitude_city": false,
+    "moderator": false,
+    "sites": false,
+    "spreadsheets": false,
+    "url_shortener": false,
+    "webmaster_tools": false,
+    "coordinate": false,
+    "coordinate_readonly": false
+  },
+  "strategy": "google-oauth2",
+  "name": "google-oauth2",
+  "is_domain_connection": false,
+  "enabled_clients": [
+    "Default App"    
+  ]
+}
