+++ custom-pages/connections/nice-waad.json
@@ -1,0 +1,31 @@
+{
+  "options": {
+    "domain_aliases": [
+      "nice.org.uk"
+    ],
+    "tenant_domain": "nice.org.uk",
+    "use_wsfed": false,
+    "useCommonEndpoint": false,
+    "client_id": "",
+    "client_secret": "",
+    "domain": "nice.org.uk",
+    "waad_protocol": "openid-connect",
+    "app_id": "",
+    "ext_groups": true,
+    "max_groups_to_retrieve": "",
+    "api_enable_users": true,
+    "waad_common_endpoint": false,
+    "basic_profile": true,
+    "ext_profile": false,
+    "ext_nested_groups": true,
+    "app_domain": "",
+    "thumbprints": [],
+    "granted": true
+  },
+  "strategy": "waad",
+  "name": "nice-waad",
+  "is_domain_connection": false,
+  "enabled_clients": [
+    "Default App"
+  ]
+}
