+++ custom-pages/database-connections/Identity/database.json
@@ -1,0 +1,37 @@
+{
+  "options": {
+    "mfa": {
+      "active": false,
+      "return_enroll_settings": true
+    },
+    "passwordPolicy": "good",
+    "disable_signup": false,
+    "requires_username": false,
+    "brute_force_protection": true,
+    "strategy_version": 2,
+    "enabledDatabaseCustomization": true,
+    "customScripts": {
+      "login": "./login.js",
+      "get_user": "./get_user.js"
+    },
+    "import_mode": true,
+    "configuration": {
+      "wtrealm": "",
+      "apikey": "",
+      "url": "",
+      "port": "443",
+      "createuserspath": "",
+      "identityapikey": "",
+      "hostname": ""
+    }
+  },
+  "strategy": "auth0",
+  "name": "Identity",
+  "is_domain_connection": false,
+  "realms": [
+    "Identity"
+  ],
+  "enabled_clients": [
+    "Default App"
+  ]
+}
