+++ custom-pages/grants/OctopusDeploy-HostedPages.json
@@ -1,0 +1,46 @@
+{
+  "client_id": "OctopusDeploy-HostedPages",
+  "audience": "",
+  "scope": [
+    "read:client_grants",
+    "create:client_grants",
+    "delete:client_grants",
+    "update:client_grants",
+    "read:clients",
+    "update:clients",
+    "delete:clients",
+    "create:clients",
+    "read:client_keys",
+    "update:client_keys",
+    "delete:client_keys",
+    "create:client_keys",
+    "read:connections",
+    "update:connections",
+    "delete:connections",
+    "create:connections",
+    "read:resource_servers",
+    "update:resource_servers",
+    "delete:resource_servers",
+    "create:resource_servers",
+    "read:rules",
+    "update:rules",
+    "delete:rules",
+    "create:rules",
+    "read:rules_configs",
+    "update:rules_configs",
+    "delete:rules_configs",
+    "read:email_provider",
+    "update:email_provider",
+    "delete:email_provider",
+    "create:email_provider",
+    "read:tenant_settings",
+    "update:tenant_settings",
+    "read:grants",
+    "delete:grants",
+    "read:guardian_factors",
+    "update:guardian_factors",
+    "read:email_templates",
+    "create:email_templates",
+    "update:email_templates"
+  ]
+}
