+++ custom-pages/rules-configs/apikey.json
@@ -1,0 +1,4 @@
+{
+  "value": "",
+  "key": "apikey"
+}
