+++ custom-pages/rules-configs/createuserspath.json
@@ -1,0 +1,4 @@
+{
+  "value": "",
+  "key": "createuserspath"
+}
