+++ custom-pages/rules-configs/hostname.json
@@ -1,0 +1,4 @@
+{
+  "value": "",
+  "key": "hostname"
+}
