+++ custom-pages/rules/Force email verification.js
@@ -1,0 +1,7 @@
+function (user, context, callback) {
+  if (!user.email_verified) {
+    return callback(new UnauthorizedError('Please verify your email before logging in.'));
+  } else {
+    return callback(null, user, context);
+  }
+}
