+++ custom-pages/rules/Force email verification.json
@@ -1,0 +1,7 @@
+{
+  "enabled": false,
+  "script": "./Force email verification.js",
+  "name": "Force email verification",
+  "order": 5,
+  "stage": "login_success"
+}
